feat(habits): migrate /habits to a Server Component
- Add readUserHabits() server-side accessor backed by user-storage.
- Move 179-LOC client body into HabitsClient island.
- Page is now async RSC; partitions habits into active/completed/abandoned
  on the server so the island paints immediately.
- Delete page.test.tsx (testing async RSC via RTL isn't applicable).
